requirements.txt
SQLAlchemy~=2.0.34
python-telegram-bot[rate-limiter]~=21.5
python-telegram-bot[webhooks]~=21.5
python-telegram-bot[job-queue]~=21.5
redis~=5.1.0b7
mysql-connector-python~=9.0.0
pytonapi==0.4.0
pytonconnect==0.3.1
qrcode==7.4.2
